if [ ! -e $ASM.prot ]

+++ genetics/euk_seq.sh
@@ -38,49 +38,51 @@ cd $DIR
 
 ls -laF $ASM
 $THIS/dna2stat $ASM  -log $LOG > $ASM.stat
-echo "GeneMark-ES 4.69" > annot_software
 
 
-# genemark.gtf
-FUNGUS_PAR=""
-if [ $FUNGUS == 1 ]; then
-  FUNGUS_PAR="--fungus"
+if [ ! -e $ASM.prot ]; then
+  echo "GeneMark-ES 4.69" > annot_software
+
+  # genemark.gtf
+  FUNGUS_PAR=""
+  if [ $FUNGUS == 1 ]; then
+    FUNGUS_PAR="--fungus"
+  fi
+  rm -rf data/
+  rm -rf info/
+  rm -rf output/
+  rm -rf run/
+  rm -f gmes.log
+  rm -f run.cfg
+  # Approximate
+  /usr/local/gmes/4.69/bin/gmes_petap.pl  --ES  $FUNGUS_PAR  --sequence $ASM  --soft_mask 0  --cores 4  &>> $LOG
+  echo -e "\nAnnotation finihsed!\n" >> $LOG
+  $THIS/../rm_all.sh data
+  rm -r info/
+  rm -r output/
+  rm -r run/
+  rm gmes.log
+  rm run.cfg
+  rm gmhmm.mod
+
+  # PAR
+  PROT_LEN=150
+
+  section "$ASM.prot"
+  $THIS/GeneMark2CDS $ASM genemark.gtf  -gtf  -gencode 1  -prot $ASM.prot  -prot_len_min 20  -ambig 10  -log $LOG  -qc
+    # was: -prot_len_min 60
+
+  section "$ASM.cds"
+  $THIS/GeneMark2CDS $ASM genemark.gtf  -gtf  -gencode 1  -cds $ASM.cds  -prot_len_min $PROT_LEN  -complete  -log $LOG  -qc
+  gzip genemark.gtf
+
+  section "$ASM.hash-CDS"
+  $THIS/fasta2hash $ASM.cds $ASM.hash-CDS  -cds  -gene_finder "GeneMark"  -prot_len_min $PROT_LEN  -log $LOG  -qc
+
+  section "$ASM.hash-PRT"
+  $THIS/fasta2hash $ASM.cds $ASM.hash-PRT  -cds  -gene_finder "GeneMark"  -prot_len_min $PROT_LEN  -log $LOG  -translate  -qc  
+  rm $ASM.cds
 fi
-rm -rf data/
-rm -rf info/
-rm -rf output/
-rm -rf run/
-rm -f gmes.log
-rm -f run.cfg
-# Approximate
-/usr/local/gmes/4.69/bin/gmes_petap.pl  --ES  $FUNGUS_PAR  --sequence $ASM  --soft_mask 0  --cores 4  &>> $LOG
-echo -e "\nAnnotation finihsed!\n" >> $LOG
-$THIS/../rm_all.sh data
-rm -r info/
-rm -r output/
-rm -r run/
-rm gmes.log
-rm run.cfg
-rm gmhmm.mod
-
-
-# PAR
-PROT_LEN=150
-
-section "$ASM.prot"
-$THIS/GeneMark2CDS $ASM genemark.gtf  -gtf  -gencode 1  -prot $ASM.prot  -prot_len_min 20  -ambig 10  -log $LOG  -qc
-  # was: -prot_len_min 60
-
-section "$ASM.cds - long"
-$THIS/GeneMark2CDS $ASM genemark.gtf  -gtf  -gencode 1  -cds $ASM.cds  -prot_len_min $PROT_LEN  -complete  -log $LOG  -qc
-gzip genemark.gtf
-
-section "$ASM.hash-CDS"
-$THIS/fasta2hash $ASM.cds $ASM.hash-CDS  -cds  -gene_finder "GeneMark"  -prot_len_min $PROT_LEN  -log $LOG  -qc
-
-section "$ASM.hash-PRT"
-$THIS/fasta2hash $ASM.cds $ASM.hash-PRT  -cds  -gene_finder "GeneMark"  -prot_len_min $PROT_LEN  -log $LOG  -translate  -qc  
-rm $ASM.cds
 
 
 if [ $PFAM ]; then
